moved fetchID to app.js
